requirements.txt
attrs==25.1.0
click==8.0.1
django-environ==0.4.5
structlog==21.1.0
valkey==6.1.0
rich==13.9.4